nvmesensor: Check SCS for secondary controllers
The Secondary Controller State indicates if the secondary controller has
been enabled by CC.EN. For disabled controllers, the StorageController
and NVMeAdmin interface should not be created.

It is because the ctrl_list will list all controllers in the system
regardless of the enable status. The number of supported controllers can
be large compare to the active ones in the system. In our example
system, the total controllers in one SSD can live up to 128 but we
only enble one or two of them. The unused controller can seriously slow
down the performance of dbus if we chose to expose them.

Signed-off-by: Hao Jiang <jianghao@google.com>
Change-Id: I6fdbb503550968d013e16822f9d11e9962bb93a4