nvmesensor: chuck fetch for telemetry logs
The telemetry logs could be larger than megabytes so fetch the whole log
at one shot could block the other transactions on the same bus.

This change slices the transaction into 4kB chucks and repush the
remaining tasks into the worker thread so other tasks (such as health
poll) can be scheduled in the middle.

Signed-off-by: Hao Jiang <jianghao@google.com>
Change-Id: I4f5e88b73cbb850954ea39fde185bb4016b8e25c