nvmesensor: add markAvailable with adaptive timer
Added markAvailable() function for NVMeSubsys. When the subsystem health
poll detects the device absence due to various reasons. The subsystem
can be marked as unavailable.

Current function during subsystem unavaibility is to reducing the
polling cadence, because an absent device can no longer report the
health data and it usually takes time for the communication/NVMe device
to recovery from absence.

Furthur task for markAvailable() will include a dbus interface to inform
the status of NVMe device (subsystem).

Signed-off-by: Hao Jiang <jianghao@google.com>
Change-Id: Ie69b0dee23685d4cf9f712894d396f975c3419b0