nvme: Rework NVMeError to throw sdbus subclasses
Generated phosphor-dbus-interfaces bindings must only throw the
exception subclasses listed in the interface definition.  Previously
NVMeError was being thrown with a name matching the known subclasses,
which worked for non-binding methods (async manually implemented ones)
but not the generated ones.

Instead we now throw the subclasses directly. This means we lose the
more descriptive text passed back from libnvme. Future modifications to
the sdbusplus exceptions could perhaps let us customise the description
text.

The descriptive text is still printed to the nvmesensor output so is
available in system logs.

Signed-off-by: Matt Johnston <matt@codeconstruct.com.au>
Change-Id: I086a831ded26639c63681a47cf97480a27e40733