nvmesensor: add primary info for plugins
Now the controller plugins can distinguish if the attached NVMe
controller is primary or secondary. Because some of the Vendor Unique
Admin cmd can only be processed via the primary controller. The
informatin now pass down to the plugin as soon as the controller
plugin is constructed so that the plugin doesn't need to rescan the
controll list for such information.

Signed-off-by: Hao Jiang <jianghao@google.com>
Change-Id: Ic5a142b89158f8a5fe23b33c4c52847d269448da