NVMeController: Expose disable() to prevent queueing jobs prior to destruction
The lifecycle of an NVMeController is partially tied to related D-Bus interfaces
where the method lambdas promote its weak pointer to shared. This interplays
with e.g. the ctemp timer determining that the related endpoint has become
inaccessible and the entire subsystem being marked as non-functional.

A call through the NVMeController instance's D-Bus APIs will hold the D-Bus
object available until the call completes, providing opportunity for further
calls to occur even though the the object is only exposed as an existing call is
in progress. These subsequent calls interfere with the effort to destroy the
controller and rescan the endpoint. Rescanning the endpoint invalidates the
nvme_mi_ctrl_t held by the NVMeController instance. We must prevent it from
occurring prior to all associated NVMeController instances being destructed else
the held nvme_mi_ctrl_t will point to an invalid object in memory.

Expose NVMeController::disable() as a trapdoor state transition: Once
NVMeController::disable() is invoked all calls through the query APIs will be
rejected and no further jobs will be submitted to the associated NVMeMi Worker
IO thread. The only exit from the disabled state is for the NVMeController to be
destructed.

Change-Id: If991588d5c17d021b49619e2bd610b65d88fca88
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>