NVMeProgress: Expose abort() for cancelling progress objects
If progress is prevented by an event outside the context of an active query the
progress object should signal an out-of-band failure to any consumers. An abort
might occur, for example, as a result of the NVMe MI endpoint becoming
unavailable prior to completing the task. Marking the progress object as aborted
instead of removing it from the bus helps provide sensible semantics through the
swordfish interface.

Change-Id: Ie962199f66f231e2a10c114fc67a508e2adccf64
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>