NVMeSubsys: Clean-up invalidated controllers, volumes and associations
If the subsystem is marked as non-functional, for example due to the endpoint
becoming unavailable, we must rescan the endpoint to re-establish its
controllers, volumes and the associations. However, prior to rescanning we must
ensure that it is safe to do so. Rescanning is not safe before all in-progress
IO jobs have completed as the IO jobs hold an nvme_mi_ctrl_t, whose lifetime is
tied to each scan period (rescanning invalidates the nvme_mi_ctrl_t).

Introduce a Status::Terminating state that is entered upon invocation of
`NVMeSubsystem::markFunctional(false)`. Status::Terminating indicates that the
subsystem must eventually enter Status::Stop, but an unknown number of
nvme_mi_ctrl object pointers remain live. We maintain the Status::Terminating
state until all active IO jobs have completed, and we enter Status::Stop from
Status::Terminating by flushing the NVMeMi worker queue with a job whose
completion transitions the subsystem to the Status::Stop state.

The fundamental invariant of the Status::Stop state is that no dangling
references to nvme_mi_ctrl objects exist, asserting it safe to (re)scan the
endpoint. To uphold it we must ensure that no further jobs can be enqueued in
the IO thread subsequent to the flush job, in addition to the flush job
completion signalling any existing queued jobs have been completed. We prevent
this by requiring the subsystem be in Status::Start for any admin queries to be
enqueued, and use NVMeController::disable() to prevent enqueuing of any
controller queries.

Having addressed the safety of any remaining nvme_mi_ctrl pointers, clear out
the subsystem's own references to the controllers, volumes and their
associations to drop the objects from D-Bus once all IO jobs holding shared
pointers complete.

Change-Id: I5315b53b375cf4fcf3bab7001d2a7055958fc552
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>