nvmesensor: dynamically load the plugin
The nvmed will dynamically load all shared libraries from a certain
folder and try to create the plugin.

The plugin library should expose a C function with type of:

createplugin_t = std::shared_ptr<NVMePlugin> (*)(
    std::shared_ptr<NVMeSubsystem> subsys, const SensorData& config)

The plugin library should return a valid shared_ptr when it has been
configured in by the subsystem, otherwise return a empty ptr.

A fake plugin is created as examples and used for unit test.

Signed-off-by: Hao Jiang <jianghao@google.com>
Change-Id: Ia1e05e7c14ead92fd7a9d3ce507fcdd8dbcdb480