Disable other sensors by default
This repo is for nvmed only, so we do not need to build other sensors by
default.

Also removing the io_uring arguments from the projects wise, since
nvmed cannot enable the io_uring and undefined the flags on the
binary/library level. The reason is if the utils library and the nvmed
(or unit test) use different io_uring flag, ubsan will report failure:

https://github.com/CodeConstruct/dbus-sensors/issues/8

Signed-off-by: Hao Jiang <jianghao@google.com>
Change-Id: I3aa9911cd7fc4980f03290f92c4bd8276a81f408