MctpEndpoint: Support mctpd's endpoint recovery
mctpd implements endpoint recovery in the form of a new method,
`.Recover`, and a new property, `.Connectivity`. `.Connectivity` takes
one of two values:

1. Available
2. Degraded

When an endpoint object is created by mctpd its `.Connectivity` member
will take the value `Available`. If a peer application on the BMC
determines that the endpoint is unresponsive it can call `.Recover` on
the endpoint's mctpd object. Invocation of `.Recover` transitions
`.Connectivity` to `Degraded`. If mctpd successfully re-establishes
connectivity with the endpoint then `.Connectivity` will transition to
`Available`. If recovery is unsuccessful the endpoint's object will be
removed by mctpd.

The signals emitted by the transitions of `.Connectivity` allow peer
applications to start or stop their communication with the endpoint as
appropriate.

More detail on the design is contained in the mctpd pull-request:

https://github.com/CodeConstruct/mctp/pull/23

Change-Id: Ie2693714fb88c78d36adcf74832f01f9e2e5553d
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>