NVMeMi: Guard relevant methods with isEndpointDegraded()
Currently there are few conditions that signal that we can't communicate
with an endpoint: nvmeEP is nullptr or isMCTPconnected() returning
false. Relevant MI command methods protect their implementation by
open-coding a check for the appropriate condition.

Isolate the open-coded tests to a single helper invoked by the command
methods to provide consistency and ergonomics at the callsites.

Change-Id: I250e5dd13514bc29de9d1cba6a149a5239d875e1
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>