NVMeMi: Flush pending operations before nvme_mi_close()
Holding `mctpMtx` serializes actions on `nvmeEP` between the main thread
and jobs posted to the queue of the NVMeMiWorker thread. Releasing
`mctpMtx` on the main thread allows jobs to progress in the worker
thread, which takes the lock over each job. On release, jobs in the
worker thread's queue may access invalid memory if `nvmeEP` is closed
under `mctpMtx` on the main thread.

Instead, defer `nvme_mi_close()` via `flushOperations()` ensure nvmeEP's
associated memory remains valid for already queued jobs.

Previously `nvme_mi_close()` via `NVMeMi::stop()` was unlikely as
we lacked support for a full NVMeMi lifecycle. With the introduction
of support for endpoint recovery we bind the NVMeMi lifetime to the
lifetime of the endpoint object exposed by `mctpd`, making its existence
much more dynamic and more likely to expose issues on destruction.

On that subject, `NVMeMi::stop()` was previously called by the NVMeMi
destructor. As we're deferring the update to member state in `stop()`,
implementing the destructor this way is no-longer valid: Attempting to
capture a `NVMeMi::shared_from_this()` reference in the job's lambda
will throw a `std::bad_weak_ptr`.

If the destructor is executing it must be the case that the last shared
reference to the NVMeMi instance has been dropped. As jobs enqueued
in the worker all capture an `NVMeMi::shared_from_this()` reference,
execution of the destructor implies that no relevant jobs remain in the
worker's queue. Further, as the destructor is running there's no need to
update other state of the instance, as the instance reaches the end of
its lifetime at the end of the execution of the destructor.

Given that, reimplement the destructor as an direct (not deferred)
invocation of `nvme_mi_close(nvmeEP)`.

Change-Id: Iadf8eefbda91293d4321d96e14e3054372d35b0f
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>