NVMeMi: Lift connectivity transitions to methods
Isolate the actions required for connectivity state transitions to their
own "event" methods. This assists with reasoning about the behavior and
correctness of the implementation.

With that philosophy in mind, we also introduce a "Configured" state to
capture the property that the network ID, EID and path variables are in
a valid state, and introduce a method to perform the transition to make
sure that property is upheld in the face of exceptions from the message
handling.

Change-Id: I1a8daf678e6a1b4275d740034a8f54d7d21a64e1
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>