NVMeMi: Pass endpoint parameters directly to epConnect()
The early implementation of NVMeMi::start() combined setup of the MCTP
endpoint for the NVMe device with setup of the NVMe MI connection.
Status::Configured was introduced into the internal state machine to
explicitly separate the reset, configured and connected phases and the
expectations on class member state in each.

The overload of start() reduces the need for Status::Configured.
Remove Status::Configured now to separate the resulting refactor of
the internal state machine from the subsequent introduction of endpoint
recovery functionality.

Change-Id: I73bfa06d2b17c6af38b596fa1b42a36d32e10b55
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>