MctpEndpoint: Introduce StaticEntity container class
StaticEntity and its specialisations exist to hold an implementation of
an upper-layer protocol used with the device, abstracting over how
presence of the device is established.

Concretely, the necessary steps to talk via the NVMe MI basic management
command and the full NVMe MI protocol over MCTP have significant
differences in polling and setup. Unify the behaviors for both under
the StaticEntity type.

Change-Id: Ib951da246309fd85406b204899cbe03ee6df445b
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>