nvmed: supprot PowerState::always in NVMeSensor
On some platform (which doesn't need warm-reboot) the PowerState
property will be PowerState::always. Here we parse the PowerState
attribute and pass it into the construction function of NVMeSensor
class.
To be consistent with the previous implementation, we only initialize
the readState of Sensor class to PowerState::always if the powerState
is PowerState::always. For the remaining powerState, we still use the
default PowerState::on for the readState of Sensor class.

Tested:
https://paste.googleplex.com/6344862406803456

Google-Bug-Id: 270233965
Change-Id: I8b2c308afd1efa64a1d5f1fdb28c87d9376b7fb0
Signed-off-by: Jinliang Wang <jinliangw@google.com>