NVMeMi: Eliminate connection state races with worker thread jobs
Iadf8eefbda9 ("NVMeMi: Flush pending operations before nvme_mi_close()")
defers the close of the endpoint connection via the worker thread.
This is to ensure that queued jobs could complete before the a
reconnection attempt forces the invalidation of any captured controller
pointers.

However, we need to also to account for preventing a subsequent
`start()` call (implemented in terms of `epConnect()`) from
modifying other essential state. Jobs queued in the worker all take a
`shared_from_this()` reference to the NVMeMi instance, and dereference
through it to access members such as nvmeEP. It's not enough to protect
the controller objects alone.

Introduce `Status::Terminating`. Invocation of `epReset()` immediately
transitions to `Status::Terminating` instead of `Status::Reset`, and
moves the transition to `Status::Reset` into the deferred job. This
prevents progress in a subsequent `epConnect()` invocation until the
deferred job has executed.

Change-Id: Ic0322eec68bf34499260f823797f1edec607a189
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>