NVMeMi: Overload start() to take network and eid
Split the NVMeMi::start() method into two parts:

1. `start()`: Calls mctpd's SetupEndpoint to configure the
   endpoint on the MCTP-capable device, and then invokes

2. `start(int network, uint8_t eid)`: Which performs actions necessary
   to bring up the NVMe MI connection used by the NVMeMi instance

Separating out `start(int network, uint8_t eid)` provides a path
for external initialization of the endpoint to provide the endpoint
parameters to the NVMeMi instance. Initialization and maintenance of
MCTP endpoints is implemented in the MctpDevice and MctpEndpoint
classes, and it's intended that we move these responsibilities out of
the NVMeMi implementation.

We retain `start()` as this is currently invoked by the power state
match callback and NVMeSensorMain, which cannot yet provide these values
to NVMeMi.

Finally, this split enables removal of the Status::Configured state,
which will be done in a subsequent patch. Status::Configured was
introduced to provide a degree of isolation between connection phases
in the NVMeMi implementation, but I think that's better achieved by the
split implemented here.

Change-Id: I43f40947cfe0beb77d9a2e0dea4a6111a9d5752b
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>