MctpEndpoint: Integrate remove behavior into MctpdDevice
9498075d1375 ("Revert "MctpEndpoint: Introduce StaticEntity container
class"") outlines the need for a different design-point than we
set out on in ac513c858a9e ("MctpEndpoint: Introduce StaticEntity
container class"). The latter commit externalised the handling of
InterfacesRemoved events from the MctpDevice and MctpEndpoint classes
to use the signal as input to resume polling via a timer invoking
MctpDevice::setup(). This felt a little untidy at the time, but
manageable.

The improved understanding of the needs that drove the revert above
triggered me to reconsider this separation. Better encapsulation
of behavior is achieved by integrating the handling of endpoint
InterfacesRemoved signals into MctpDevice. MctpDevice is then
self-contained with respect to managing the lifecycle of its
MctpEndpoint. From here we can integrate handling of MctpDevice into
NVMeSensorMain such that its MctpEndpoint's remove event callback
triggers the required polling.

Change-Id: I7d0d69d7235c19ebfbdef83fe1094719fd82f2fd
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>