Revert "MctpEndpoint: Introduce StaticEntity container class"
This reverts commit ac513c858a9efaf28f040c641bdd287d6f942221.

StaticEntity was motivated by the expectation that the lifetime of an
NVMeSubsystem instance would be bounded by the lifetime of its
associated MCTP endpoint. However, that design point is contrary to
consumer needs as inventory data is exposed by the NVMeSubsystem
implementation.

Instead, the lifetime of an NVMeSubsystem instance must remain related
to the presence of its EM configuration (which is in-turn bounded by the
presence of the FRU EEPROM describing it). The lifecycle of its MCTP
endpoint should drive start() and stop() events on the NVMeSubsystem
instance instead. This ensures that inventory data exposed by the
NVMeSubsystem implementation remains available for query even in the
absence of the endpoint with which the subsystem communicates.

Change-Id: I9cd560d04de1825327cddaafa3818fee7b42634e
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>