nvmesensor: Implement MCTP endpoint recovery
MCTP endpoint recovery allows for device removals and partial or
complete failures of MCTP endpoints to be managed effectively. The
recovery implementation rests on the `.Recover` method and related
behaviors exposed by mctpd[1].

The implementation comes in several parts:

1. Adjusting NVMeMi::start() to take an MctpEndpoint

2. A new request method exposed by NVMeMi: recover()

3. NVMeSubsystem issuing a call to NVMeMi::recover() when it judges the
   endpoint to be in an error state

4. Integration of endpoint lifecycle management using events exposed by
   the MctpEndpoint instance associated with an MCTP-capable device.

Track related MctpDevice, NVMeMiIntf and NVMeSubsystem instances
in an NVMeDevice struct held by the NVMeMap type. On exposure of a
new drive configuration from EntityManager we hook the NVMeMi and
NVMeSubsystem start() and stop() calls into the MctpEndpoint lifecycle.

NVMeSubsystem's call to NVMeMi::recover() when successive NVMe MI
commands fail results in one of:

1. A determination that the endpoint is available

2. A determination that the endpoint was temporarily unreachable, but
   was reconfigured successfully

3. A determination that the endpooint cannot be recovered, and its
   representation is removed from DBus

These lifecycle events are delivered through the callbacks passed into
MctpEndpoint::subscribe().

NVMeMi::recover() needs access to the MctpEndpoint instance on which it
should invoke the `.Recover` DBus method. To accommodate that we
switch NVMeMi::start(int network, uint8_t eid) to instead take a
std::shared_ptr<MctpEndpoint> directly. Further, as we now have the
NVMeMi::recover() capability, we rework the power state match callback
to function in terms of recover() rather than start() and stop()
directly. This enables removal of NVMeMi::start(), as the endpoint
setup is now handled by the MctpEndpoint lifecycle events and
MctpDevice::setup().

With access to a std::shared_ptr<MctpEndpoint> there's now no need
for the `nid` and `eid` class members as they are encapsulated by the
MctpEndpoint. Remove those, as well as the `bus` and `addr` members,
which are encapsulated by std::shared_ptr<MctpDevice> passed through
the constructor. The removal of these members drives a cleanup of
all the logging that exploited them, which is reduced to calls to
`device->describe()` or `endpoint->describe()` as required. Finally, the
DBus call for configuring the MCTP device MTU is also impacted by the
removal of the `nid` and `eid` members. However, its functionality is
encapsulated in the MctpEndpoint implementation, so the impacted method
is converted to the MctpEndpoint implementation instead.

[1]: https://github.com/CodeConstruct/mctp/blob/69ed224ff9b5206ca7f3a5e047a9da61377d2ca7/docs/endpoint-recovery.md

Change-Id: I62c601c7b92a007f8467d36fbeecd7451fbc111d
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>