NVMeController: add States of Enabled/Disabled
Using the state variable instead of nvmeCtrl ptr as indication for
controller state.

The most important reason is the nvmeCtrl is retrieved from controller
list (via MI data structure or NVMe Admin). That controller list is
retrievable as long as the subsystem is functional, regardless of the
actual controller status (online/offline/enable/shutdown). That is why
the nvmeCtrl ptr allocated in the constructor of NVMeController class
and binding its lifetime to the class.

Thus we need an additional state to indicate controller's functionality
life cycle.

Currently, the state machine has:
  * Enabled: fully initialized and ready to serve the external client
    from DBus.
  * Disabled: detected from subsystem but functionally not ready to
    serve.

As for now, the controller is only set to Disabled during the shutdown
process, which is synchronized to the previous behavior.

Soon, it will also been used for the initialization process to prevent
the anxious clients from early interfere.

Google-Bug-Id: 326355507
Signed-off-by: Hao Jiang <jianghao@google.com>
Change-Id: I1afd8e4e1bf1561fde1ae8957668fdf3511c21f0