Unitest: Config the polling rate to 100ms
The unitest has been expand to the edge of 30s. The test was based on
the normal NVMe susystem health polling rate of 1 second. This change
will increase the polling rate by 10x to reduce the unitesting time.

When running via valgrind, remains the normal polling rate = 1s. That is
because the valgrind will serviously slow down the processing time so
that the polling rate will be bottlenecked by the handling(processing)
speed.

Signed-off-by: Hao Jiang <jianghao@google.com>
Change-Id: Ic38fe8d71b5c9fcbfb294c3ff8ae678131fd4f58