NVMeMain: stop susbsystem start/stop with MCTP status change
Currently the thermal sensors and plugs are binded to the subsystem
start/stop lifetime. With the MCTP recoverability, it binds the
start/stop also to the MCTP interface.

It is problematic since during the MCTP down period, the system still
needs the thermal and the plugin interface for functionilty like thermal
control.

Remove the subsystem.start()/stop() from the MCTP config callback, so
the subsystem only starts when the EEPROM interface detected, and stops
when the EEPROM interface removes. This behavior is align with the
previous ahead of MCTP recoverability.

Google-Bug-Id: b/333946719
Tested: http://sponge2/a4a7f497-2d94-4798-a055-b1f48c091221
Signed-off-by: Hao Jiang <jianghao@google.com>
Change-Id: Icd1d7218bbaed2b2d24ab05c44afa5c42c33723f