tests: setup: Use a consistent path to bash
The `run_command()` directive in `test/meson.build` used `/usr/bin/bash`
while `tests/setup.sh` used `/bin/bash` in the shebang line. With the
whole usr-merge thing, everything should be in `/usr/bin` going forward,
and the fact that it is the `run_command()` directive in `tests/
meson.build` that is executed suggests `/usr/bin/bash` is reasonable.

Change-Id: I3d2b444ac320ac7b13b3348ab9eb642dc0050b53
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>