tests: nvme: Lift constraint requiring the OpenBMC CI container
By exploiting dbus-run-session[1] we can isolate the behavior of the
mapper (as required for the tests) from the user's session bus. This
works regardless of whether we're running inside or outside of the CI
container.

The change works in several stages: We first lift the test_nvme
executable target out of the test definition in order to rework the
test definition to be in terms of dbus-run-session. Next, the test
script itself is reworked such that its execution bounds the execution
of the test_nvme binary. In this way the test DBus session is terminated
along with the tests. The paths to the mapper and test_nvme binaries are
passed as arguments into the script, and the tests are invoked once the
mapper instance has been set up.

A small work-around regarding the DBus environment variables is required
to meet the requirements inside the CI container. The constraints are
documented alongside the work-around.

[1]: https://manpages.debian.org/bookworm/dbus-daemon/dbus-run-session.1.en.html

Change-Id: I37b017d83b3f121f895d5994080e6ab3de4931c8
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>