NVMeSensorMain: Split out NVMeDevice
NVMeDevice was defined as a type to hold references to the MI and
subsystem instances across the lifetime of the device configuration
provided by EM.

setupMctpDevice() was a static function associated with these objects by
way of taking them as parameters. A small trick here was the conversion
of strong to weak pointers by way of the types specified for the formal
parameters. The conversion enables the desired lifetime management of
the associated DBus objects, tying them to the EM configuration (and not
the lifetime of `mctpd` DBus objects).

Really though, setupMctpDevice() could be better expressed as a method
on NVMeDevice: This way we can handle any smart-pointer type conversions
internal to the class' implementation, and applying this perspective
increases the testability of the code. A final benefit is that
by isolating the code we can refactor it to add features without
significantly impacting its readability. Ultimately the goal is to
insert a delay via a timer to mitigate stalling of the main thread due
to recovery-related event-cycles.

Here we purely do the code-motion that moves the implementation of
setupMctpEndpoint() out of NVMeSensorMain.cpp, along with defining a
method that wraps its invocation.

Change-Id: I498c43561dee61462e049b9ad43c987e1754ba0c
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>