NVMeSensorMain: Merge interface and subsystem config loops
The startup state machine of NVMeMi is (eventually) asynchronous,
and the events driving it are deferred beyond the startup of the
NVMeSubsystem. The comment justifying the split-loop approach no-longer
seems to make sense, and even if it did at some stage, the delay implied
is quite uncontrolled. Further, the NVMeMi instance will (now) reject
attempts to communicate with the endpoint until its own initialization
process is complete, meaning there will not be any additional bus
contention from subsystem-related traffic.

Change-Id: Id7754d8e45932b7f7db38cce668a8d143c7eea22
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>