NVMeDevice: Remove redundant start
The final act of starting the underlying NVMeMiIntf instance in
NVMeDevice::finalize() is redundant. The contract of
MctpEndpoint::subscribe() is that at least one of the callbacks will
be invoked at least once, to ensure visibility of the current endpoint
state. As such, if the endpoint is in the 'Available' state then
NVMeDevice::available() will be invoked, which will in-turn invoke
NVMeMiIntf::start(). Removal of the redundant call also removes the
initiation of one endpoint recovery request in the event that
that the endpoint is 'Degraded' at the time of the call to
NVMeDevice::finalize().

It's not an error to invoke NVMeMiIntf::start() multiple times, so this
hasn't been a concern, but it is unnecessary.

Change-Id: Id2c2e5af864a2fde64205b62518df68d82783796
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>