NVMeDevice: Clean up lambdas and lifetimes
setupMctpDevice() implemented the MctpEndpoint::subscribe() callbacks
entirely with in-line lambda functions.

General consensus is that long and/or complex lambdas are not
desirable[1].

Given that we're only going to add to the complexity to mitigate MCTP
endpoint recovery event cycles, unpick the lambdas by refactoring them
to call back into private methods of NVMeDevice.

Together with untangling the lambda implementations, we can
significantly simplify the captures. We make NVMeDevice inherit
std::enable_shared_from_this, after which the lambdas only require a
weak pointer to NVMeDevice.

From there we provide the usual NVMeDevice::create() overloads to
require its instantion be wrapped in a std::shared_ptr.

[1]: https://github.com/openbmc/docs/blob/master/anti-patterns.md#very-long-lambda-callbacks

Change-Id: I53574da34448355eff0aea18cda744dc72e20849
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>