NVMeMI: Remove the async timer to optimize EP
Also,
* Move the get/set timeout inside the worker thread
* Use device instead of Endpoint in error conditions

Previously, we were using the async timer to avoid the tight loop
while optimizing the MCTP endpoint. With MCTP recoverability, this
will be handle by the calls to the MCTPd to recover the endpoint.
This will also avoid the case where the handler is called when the
MCTP state is already transition to Reset resulting in segmentation
fault. Also, move the get/set timeout inside the worker threads and
guard the endpoint

Google-Bug-Id: 341322615

Change-Id: Ic3331e2b587af148b6183a9dabdf1c8f9cd547c6
Signed-off-by: Muhammad Usama <muhammadusama@google.com>