NVMe: remove the self capture for the worker thread
We found an uncaught std::system_error w/ type of
resource_deadlock_would_occur and led to a sigabrt.

It was because when the lifetime of worker is held by the worker thread
itself, the deconstruction may happen in the worker thread instead of
the main thread, leading to the deadlock.

The bug was introduced by the change
https://gbmc-review.git.corp.google.com/c/dbus-sensors/+/16062, when
removing the mutex, which protected scene from happening.

Given the mutex is no longer needed, the self capture is unnecessary
either. So the worker thread will not capture any resources.
These resources captured by the callback functions are executed in
the main thread instead of the worker thread by design, which have
been/will be avoiding this deadlock.

Google-Bug-Id: b/372694841
Signed-off-by: Hao Jiang <jianghao@google.com>
Change-Id: Ic12ab0781c612c83236ec7a987fd0ee662cc882b