Fix throwing NVMeSdBusPlusError
Previously this was inconsistent throwing nvme_ex_ptr or NVMeSdBusPlusError.
With this change it will always throw NVmeSdBusPlusError, catching a
reference.

"bugprone-throw-keyword-missing" is silenced since it seems to be a
false positive - it is a copy of a shared_ptr, not constructing a new
exception instance. Regardless, the intent of NVMeError _is_ to hold a
sdbusplus::exception_t without immediately throwing, in order to pass
errors out of asio async lambdas.

Other clang-tidy reports in NVMeError are also fixed. std::move is added
to some throw sites to avoid a "cert-err09-cpp" report.

Signed-off-by: Matt Johnston <matt@codeconstruct.com.au>