mctp: MCTPEndpoint: Introduce MCTPDevice::id()
In upcoming changes we wish to track endpoint state for each device.

The design choice for doing so is to maintain an association between
the device and its state in MCTPReactor, by contrast to embedding the
state in the MCTPDevice implementation. The reason for the former over
the latter is the choice of implementing a Mealy FSM to manage events
that the reactor is responsible for handling; encapsulating the state
in the MCTPDevice implementation splits the reactors responsibilities
between the reactor and device implementations, which is undesirable.

The association will be maintained in a std::map. The tidy way to do
this would be to maintain std::weak_ptrs to MCTPDevice instances as
keys, but we're unable to do so due to library constraints.

Instead, introduce MCTPDevice::id(), which will be used to key
the endpoint state map in place of weak_ptrs. MCTPDDevice::id() is
implemented in terms of std::hash over relevant members. However,
one of the members is std::vector<uint8_t>, for which no std::hash
specialization exists.

Solve that latter problem by introducing an implementation of
Fowler-Noll-Vo hashing by suggestion of the example implementation of
std::hash::operator() from cppreference.com[1], for both ILP32 and LP64
platforms, and using that to hash std::vector<uint8_t>. The result is
then folded into the subsequent value returned by MCTPDDevice::id().

[1]: https://en.cppreference.com/w/cpp/utility/hash/operator().html

Change-Id: I807f922e58dd086e310e2433df469c6337acc04b
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>